move tutorial-related codes to TutorialUtils.ts
